+++ demos/demo-factories/tests/test_demo_factories/test_demo_factories.py
@@ -1,8 +1,8 @@
 from unittest import IsolatedAsyncioTestCase
 
-import demo_registrydao_hash
+import demo_factories
 
 
 class ExampleTest(IsolatedAsyncioTestCase):
     async def test_example(self):
-        assert demo_registrydao_hash
+        assert demo_factories